fix 1 пункт

+++ hw04/part1.py
@@ -1,6 +1,6 @@
 from sys import argv
 
 if len(argv) == 4:
-    print(int(argv[1])*int(argv[2]) + int(argv[3]))
+    print(float(argv[1]) * float(argv[2]) + float(argv[3]))
 else:
     print("Ошибка ввода данных")
